+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 
 CHANGES:
 
-* Added a new script list-all-beanstalks.py. ([@TGWolf][])
+* Added a new section for IAM scripts. ([@TGWolf][])
+* Added a new section for EC2 scripts. ([@TGWolf][])
+* Added a new script elasticbeanstalk/list-beanstalks.py. ([@TGWolf][])
+* Added a new script iam/list-users.py script. ([@TGWolf][])
+* Added a new script iam/list-access-keys.py script. ([@TGWolf][])
+* Added a new script ec2/list-key-pairs.py script. ([@TGWolf][])
 * Added better error detection to all scripts. ([@TGWolf][])
 
 ## 1.0.2 (November 6, 2018)